+++ src/utils/chunkArray.ts
@@ -1,0 +1,10 @@
+export function chunkArray(array: any[], chunkSize: number): any[][] {
+  const result = [];
+
+  for (let i = 0; i < array.length; i += chunkSize) {
+    const chunk = array.slice(i, i + chunkSize);
+    result.push(chunk);
+  }
+
+  return result;
+}
